requirements.txt
praw==7.7.1 
pymongo==4.0.1
python-dotenv==0.20.0